requirements-cloud.txt
# Core ML dependencies
numpy==1.24.3
pandas==2.2.3
torchvision==0.21.0
Pillow==11.1.0
huggingface-hub==0.28.1
deepdanbooru==1.0.3
onnxruntime==1.16.3
opencv-python==4.11.0.86

# Cloud Function specific dependencies
functions-framework==3.5.0
google-cloud-storage==2.16.0
requests==2.32.3

# Qdrant dependencies
qdrant-client==1.7.0